Added comment depth and friendly formattig for GET /quotes endpoint. Finished the REST Services for DELETE and POST requests
Firstly, added depth to the step-by-step procedural comments in the "GET /quotes" endpoint. Rememeber, within the GET /quotes, we have a conditional set up to check for any queries for year, author or id in the URI.

The, we completed two more mandatory endpoints: DELETE and POST requests on /quotes.
1. To delete an entry, we should make sure that we delete exactly the entry we want - this means that we need to identify it by a unique identifier (hint: what is the one field that gives each record in the db a unique value?)

2. Finally, we learn how to convert POST requests in to records in our Quotes table! First, we want to make sure we have enabled the body-parser module. (Let's recall back in our lesson for the exact formatting needed when passing request body values in to Node via Express)